requirements.txt
pydriller==2.1.3
python-dotenv==1.0.0
gitpython>=3.1.40